Reference

GhilSellersEBM.ContinousGhilSellersParameters

GhilSellersEBM.Grid

GhilSellersEBM.NeumannFD

GhilSellersEBM.D

GhilSellersEBM.R_i

GhilSellersEBM.R_o

GhilSellersEBM.c

GhilSellersEBM.g

GhilSellersEBM.ghilsellers_ebm!

GhilSellersEBM.k

GhilSellersEBM.α

GhilSellersEBM.ContinousGhilSellersParameters — Type

ContinousGhilSellersParameters{T}

Holds all parameters needed for the Ghil Sellers 1D EBM. The model is given in CGS units. Uses the tabled values from the Ghil Paper and interpolates on a new grid.

Initialization

ContinousGhilSellersParameters(g::Grid)

g::Grid{T}

ϕ::AbstractVector{T} latitude vector

∂ₓ::NeumannFD{T} FD scheme

T_0::AbstractVector{T} initial condition given in the Ghil paper

C::AbstractVector{T} heat capacity

Q::AbstractVector{T} solar irradiance

b::AbstractVector{T} empirical albedo coefficient

z::AbstractVector{T} average surface elevation

k_1::AbstractVector{T} sensible heat flux coefficient

k_2::AbstractVector{T}coeffcient of latent heat flux in the atmosphere

c_1::T = 0.009 empirical albedo coefficient

c_2::T = 0.0065 empirical albedo coefficient

c_3::T = 1.9e-15 empirical emissivity coeffcient

c_4::T = 6.105*0.75*exp(19.6) latent heat flux coefficient

c_5::T = 5350 latent heat flux coefficient

σ::T = 1.356e-12 Stefan Boltzmann constant (in CGS system)

m::T = 0.5 Atmospheric attenuation coefficient (0.5 for present conditions)

T_m::T = 283.16 reference temperature

GhilSellersEBM.Grid — Type

Grid{T}

1D discretization Grid

Initialization

Grid(x::AbstractRange{T})

x: coordinate axis

Fields

N: Number of grid points

x: Coordinate axis

Δx: Spacing between grid points

GhilSellersEBM.NeumannFD — Type

NeumannFD{T}

1D - Finite Difference Scheme matrix with Neumann boundary conditions

Initialization

NeumannFD(T::DataType, n::Integer, Δx::Number=1)

T: precision used (e.g. Float64)

n: number of grid points

Δx: spacing between grid points

NeumannFD(grid::Grid{T})

Usage

g = Grid(1:1:10)
∂ₓ = NeumannFD(g)
dfdx = ∂ₓ(f)

GhilSellersEBM.D — Method

D(T,p::ContinousGhilSellersParameters)

1D Diffusion with Neumann boundary conditions applied

GhilSellersEBM.R_i — Method

R_i(T,p::ContinousGhilSellersParameters)

Incoming radiation at discretized coordinate x and temperature T

GhilSellersEBM.R_o — Method

R_o(T,p::ContinousGhilSellersParameters)

Outgoing radiation according to Stefan-Boltzmann law with emissivity reduced to account for greenhouse gases

GhilSellersEBM.c — Method

c(T,p::ContinousGhilSellersParameters)

Emissivity coefficient accounting for greenhouse gases

GhilSellersEBM.g — Method

g(T,p::ContinousGhilSellersParameters)

Latent heat flux contribution

GhilSellersEBM.ghilsellers_ebm! — Method

ghilsellers_ebm!(du,u,p,t)

RHS of the PDE, to be used with DifferentialEquations.jl

GhilSellersEBM.k — Method

k(T,p::ContinousGhilSellersParameters)

Heat flux, first term sensible heat flux, second term latent heat flux

GhilSellersEBM.α — Method

α(T,p::ContinousGhilSellersParameters)

Albedo at discretized coordinate x and temperature T. Albedo is cutoff at minimum 0.25 and maximum 0.85