Reference

GhilSellersEBM.ContinousGhilSellersParameters

GhilSellersEBM.Grid

GhilSellersEBM.NeumannFD

GhilSellersEBM.D

GhilSellersEBM.D

GhilSellersEBM.R_i

GhilSellersEBM.R_i

GhilSellersEBM.R_o

GhilSellersEBM.R_o

GhilSellersEBM.c

GhilSellersEBM.c

GhilSellersEBM.g

GhilSellersEBM.g

GhilSellersEBM.ghilsellers_ebm!

GhilSellersEBM.k

GhilSellersEBM.k

GhilSellersEBM.α

GhilSellersEBM.α

GhilSellersEBM.D — Method

D(T,p::ContinousGhilSellersParameters)

1D Diffusion with Neumann boundary conditions applied

GhilSellersEBM.R_i — Method

R_i(T,p::ContinousGhilSellersParameters)

Incoming radiation at discretized coordinate x and temperature T

GhilSellersEBM.R_o — Method

R_o(T,p::ContinousGhilSellersParameters)

Outgoing radiation according to Stefan-Boltzmann law with emissivity reduced to account for greenhouse gases

GhilSellersEBM.c — Method

c(T,p::ContinousGhilSellersParameters)

Emissivity coefficient accounting for greenhouse gases

GhilSellersEBM.g — Method

g(T,p::ContinousGhilSellersParameters)

Latent heat flux contribution

GhilSellersEBM.ghilsellers_ebm! — Method

ghilsellers_ebm!(du,u,p,t)

RHS of the PDE, to be used with DifferentialEquations.jl

GhilSellersEBM.k — Method

k(T,p::ContinousGhilSellersParameters)

Heat flux, first term sensible heat flux, second term latent heat flux

GhilSellersEBM.α — Method

α(T,p::ContinousGhilSellersParameters)

Albedo at discretized coordinate x and temperature T. Albedo is cutoff at minimum 0.25 and maximum 0.85